requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pyyaml>=6.0
requests>=2.31.0
